Created Function_definition_and_call.py
Based on function definition and call 
Write functions for addition, subtraction, multiplication and division of all numbers from a list and print the final results of each function. In your main program, ask for number of elements, and actual elements of the list. Pass the list to each function call of arithmetic operations.