Update greatest of three numbers.py
Hey, I just try to add some more value to the program by handling the ValueError as if someone give a value that is not integer or float than that Error in red colour don't seem good. So, handling it is a good practice. 
Please review it and if it's ok than merge it.